02-backend-app-statefulset/api-fastapi/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
httpx
factory_boy